[PREMANIEMENT] Extrais `ElementsConstructibles` de `ListeItems`
