+++ src/modeles/entitesExternes.js
@@ -1,7 +1,7 @@
 const EntiteExterne = require('./entiteExterne');
-const ListeItems = require('./listeItems');
+const ElementsConstructibles = require('./elementsConstructibles');
 
-class EntitesExternes extends ListeItems {
+class EntitesExternes extends ElementsConstructibles {
   constructor(donnees) {
     super(EntiteExterne, { items: donnees.entitesExternes });
   }
