+++ src/modeles/listeItems.js
@@ -1,13 +1,6 @@
 const InformationsHomologation = require('./informationsHomologation');
 
 class ListeItems extends InformationsHomologation {
-  constructor(ClasseItem, donnees, referentiel) {
-    super();
-    const { items } = donnees;
-    this.items = items.map((donneesItem) => new ClasseItem(donneesItem, referentiel));
-    this.referentiel = referentiel;
-  }
-
   item(index) {
     return this.items[index];
   }
