+++ src/modeles/mesuresSpecifiques.js
@@ -1,8 +1,8 @@
-const ListeItems = require('./listeItems');
+const ElementsConstructibles = require('./elementsConstructibles');
 const MesureSpecifique = require('./mesureSpecifique');
 const Referentiel = require('../referentiel');
 
-class MesuresSpecifiques extends ListeItems {
+class MesuresSpecifiques extends ElementsConstructibles {
   constructor(donnees = {}, referentiel = Referentiel.creeReferentielVide()) {
     const { mesuresSpecifiques } = donnees;
     super(MesureSpecifique, { items: mesuresSpecifiques }, referentiel);
