+++ src/modeles/partiesPrenantes/partiesPrenantes.js
@@ -1,7 +1,7 @@
-const ListeItems = require('../listeItems');
+const ElementsConstructibles = require('../elementsConstructibles');
 const PartiePrenante = require('./partiePrenante');
 
-class PartiesPrenantes extends ListeItems {
+class PartiesPrenantes extends ElementsConstructibles {
   constructor(donnees = {}) {
     const { partiesPrenantes = [] } = donnees;
     super(PartiePrenante, { items: partiesPrenantes });
